+++ src/components/hyphen.component.jsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import { theme } from '@opuscapita/oc-cm-common-layouts';
 
 const HyphenSpan = styled.span`
-  padding: 0 ${theme.gutterWidth} 8px ${theme.gutterWidth};
-  align-self: flex-end;
+  padding: 30px ${theme.gutterWidth} 8px ${theme.gutterWidth};
+  align-self: flex-start;
   margin-bottom: calc(2 * ${theme.gutterWidth});
 `;