+++ src/translations/default-props.js
@@ -49,4 +49,5 @@ export default {
     plural: 'weekdays',
     singular: 'weekday',
   },
+  ok: 'OK',
 };